Refactoring (#27)
* Long mediaId, movieId to long

* Double popularity to BigDecimal

+++ api-tests/src/main/java/org/themoviedb/models/movie/RatedMovieDto.java
@@ -14,10 +14,10 @@ import java.math.BigDecimal;
 public class RatedMovieDto extends MovieDto {
     private final BigDecimal rating;
 
-    public RatedMovieDto(final Long id,
+    public RatedMovieDto(final long id,
                          final String title,
                          final String originalTitle,
-                         final Double popularity,
+                         final BigDecimal popularity,
                          final BigDecimal rating) {
         super(id, title, originalTitle, popularity);
         this.rating = rating;
